requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
bcrypt
cryptography
pandas
Flask-Mail 